Refactor `Import::ActualHistory` into a service
Refactor `Import::ActualHistory` from a model into a service, namespaced
under the existing `actual` structure.

NB: I’m not sure what to do with the `Import::RowError` class - is it
okay where it is, is it worth folding it into the service, or would it
be better to see if we can reuse it within other import services?