Forecast.set_value must return a forecast or nil
Ensure `Forecast.set_value` always returns a forecast or nil. The return
value of this method is used by other methods, which rely on it always
returning a `Forecast` instance or nil.

Previously, when recording a historical event, such as when a Partner
Org user uploaded a revision to a forecast that was already part of an
approved report, the value returned was the value of the historical
event instead of the latest forecast. This caused erroneous results
being shown to users.