Include refunds and adjustments in activities' total spend
Previously the total spend equalled the sum of actuals for an activity and its
descendants. However, in the UI there was already a sum of actuals. Furthermore,
users were having to download data and do their own calculations to work out the
combined sum of all actuals, refunds and adjustments. Therefore this updates the
total spend figure to use this calculation instead