Logout after scenarios authenticating a user
The test suite has a number of tests that can fail depending on the
order in which they are run.

We identified one frequent culprit in scenarios that perform user
authentication but never log the users out.

We’ve now added logouts after each scenario where a user is logged in.