Disallow inactive implementing organisations
Previously organisations could be set as inactive, but they would still appear in
the list of possible implementing organisations, and could successfully be
registered to an activity as an implementing organisation. This removes inactive
organisations from the list of possible implementing organisations. It also updates
the model and the controller's `create` action to add additional assurance against
registering inactive organisations (model validation and a branching user journey
depending on whether the implementing organisation was successfully added)