Separate comments in report CSVs with a pipe
This change was done at the request of the BEIS team who handles
comments in reports. In order to enable their workflow, comments are now
separated with a pipe character (`|`) instead of newlines and dashes.

Potential future changes might address including the type of comment
(such as “Refund” or “Adjustment”).