Make incomplete parent activity error more explicit
Users have reported being unable to add Level C and D activities to some
Level B activities, as they were met with the error message: "The parent
activity is not valid". The error was caused by the Level B activity not
being in the correct `form_state` (it must be `completed`).

This message doesn't feel clear enough, as it doesn't tell the user how
to rectify the issue - stating that the parent is actually "incomplete".

Hopefully now this should make it clearer to the user that they have to
go in and finish creating the Level B activity.