Rename slightly misleading private method
When trying to investigate a reported bug, it wasn't clear to me that
`fetch_parent` would be responsible for setting errors too.

I've gone back and forth on refactoring this out into a private method
just responsible for validating the parent activity once it's been
fetched, but it felt a bit messy to me living in the `initialize`
method. I'd like to refactor this in future, but it feels like more
involved job that might re-write a lot of the `ActivityCreator` logic
and I don't want to spend more time on it now.