Use a helper to determine if XML download shown
We use this logic in multiple places, which makes it hard to change if
the business logic ever changes.

The logic was also slightly incorrect, in that it didn’t check for
permissions separately for projects and third-party projects. While this
is how policies are defined currently, with similar permissions for both
levels, it might not always be the case.