Set funds instance var in reports controller
This refactor will make it easier in the future to change which funds
are used to generate the fund dropdown in the new report form, by making
this a responsibility of the controller instead of the view.

This will make it easier to hide a specific fund with feature flagging.