Add friendly methods to check feature flag status
While the method to check whether a given feature is active for a given
user is intuitive, the way of checking whether a given feature is active
for a given user group is more contorted.

Our use case is that we’ll only have one feature for the foreseeable
future, so in order to make the code more readable and easier to
maintain, I’ve defined these syntactic sugary methods to use across
the app instead of peppering it with `ROLLOUT` calls.

I’ve also moved stubbing `ROLLOUT` to the Rails helper, so all scenarios
that use the helper can start with a clean feature flag slate.