Update `db/schema` for latest Postgres version
This changes every time I run `db:migrate` on Postgres version `14.5`,
and is muddied when adding a new field.

This seems to add the `postgis` extension here, which is used by the
`pg_restore` part of `script/db-restore` and may fix certain issues with
it.

I'd like to verify this before merging to `develop`.