Show ODA/Non-ODA on activity details
Show the ODA/Non-ODA status on activity details.

Show “Add”/“Edit” link if the user has permissions to edit the activity.

The method that decides between “add” and “edit” was using `blank?` to
check for a value being present. This test isn’t accurate for booleans,
where `false` is a valid value. We’re replacing `blank?` with `nil?`,
which is more accurate for checking whether a value has been filled in
at all.