Commit `db/schema` to codebase
This was generated with Postgres v14.5. It's been changing for a while,
and I'd like to check this into the main codebase. We're due to add some
new fields for ISPF and I'd like to not muddy the water with these
changes.

The one thing I'm a bit confused about here is the removal of the index
on `email` which happened automatically.

Co-authored-by: George <george@dxw.com>