Authorise new/create against activity edit policy
Adding an implementing organisation to an activity counts as an edit /
update operation on the activity.

Using the generic `:project` policy was ignoring the existence or not of
an editable report, which is taken into consideration when authorising
edit/update actions on a specific activity.

Checking the `create` permission, while having functionally the same
result, isn’t semantically correct, and the results could change
unexpectedly if the permissions diverge.