Properly authorise showing Remove impl org button
Using the generic `:project` policy doesn’t take into account the
existence or not of an editable report.

Removing an implementing org is authorised in the controller against the
specific activity instance, and not permitted if there isn’t an editable
report for the partner org and fund.

The difference in checks leads to inconsistent behaviour, such as
being shown the Remove button but the action not being permitted, and
resulting in an authorisation error being shown to the user.