Hide fields from activity details based on `requires_X?` methods
This uses all the new and existing `requires_X?` instance methods in the
`Activity` class to determine whether to show fields in the activity details
tab/page. One or two of these were already being used, while on some other
fields the old logic of the `requires_X?` method was simply being replicated in
the view - this brings the logic up to date and keeps the code DRY

It also updates the logic for displaying SDGs, however this does not use a
`requires_X?` method. Since the validation of `sdg_1` is based on `sdgs_apply?`
rather than a `requires_X?` method, the logic as to whether to skip this form
step and show it in the view are kept separate/plain to avoid confusion about
the difference between `sdgs_apply?` and `requires_sdgs?`

The country partner organisations are left in for now - there are some activities
for which they are not required but which still have them, and previously these would
have been shown. Hiding them based on `requires_country_partner_organisations?`
would also cause a few tests to fail