Set default currency & transaction_type
on all classes that inherit from `Transaction`.

Previously, this was only done for `Readjustment`s and `Refund`s, but
`Actuals` are also a type of `Transaction` and are exported in IATI XML.

Because `Actual`s didn't have a default value for `currency`, this meant
it was outputting XML tags with an empty string value e.g.

```
<transaction>
<transaction-type code="3"/>
<value currency="" value-date="2022-03-31">1000</value>
```

I'm not sure why the default value wasn't added for Actuals
[here](https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1383),
it may have been an oversight, but as we export all
`Transaction`-inheritors to the XML, I thought it would be safest to put
the default values on the `Transaction` class itself so we don't forget
this in future.

Although the `transaction_type` doesn't appear to be an issue for IATI
XML (as far as I'm aware) it feels best to set this to the default value
("3") as this seems to be the only real value we set it to in the
codebase.

We'll want to re-run the data migration to backfill these values:
`db/data/20211004_backfill_missing_attrs_on_refunds_and_adjustments.rb`
on production to fix this.