Refactor activity CSV column headings into YAML file
This refactors the management of activity CSV column headings into a YAML file.
Previously we had two methods, `column_headings` and `level_b_column_headings`
in `Activity::Import`, which were adding and subtracting columns to a base
`FIELDS` constant. This already had a bit of a smell to it, but with the
upcoming CSV templates for ISPF, four more methods would have had to be added.
Refactoring the management of this out into a YAML file makes it easier to
read and maintain

We've opted to read the files in an initializer and store them as a
global variable. Some people have strong anti-global-variable opinions,
but we think this is safe for our purposes. We believe it's an
improvement over memoizing with a class variable (not thread-safe) and
re-reading from the file every time `csv_columns` is called.

Two new methods have been included in the `Activity::Import` class to import and
filter the YAML data

Co-authored-by: George <george@dxw.com>