Don't ignore CSV fixtures
Previously all CSVs were ignored, including those in the fixtures folder (unless
you'd force staged them in the past). All XMLs were ignored, except those in the
fixtures folder. The latter should be the case for CSV fixtures too, since these
are test dependencies. This brings them under the purview of Git, and further
out of the purview of the developer's memory