Use activity upload form partial in new views
Previously the activity upload form partial was made compatible with new views,
so this updates the activity new views to use this

A `show_hint` local has been added so that the upload input hint can still be
shown at Level C/D. If the new Level C/D activity view ends up following the
same structure as Level B - a form for each fund - this will be able to be
removed. However, it may be possible to keep the Level C/D new view scoped to a
specific fund, based on the report?