Ensure linked activities reciprocity
The first step to linking activities: ensure that when we link an
activity to another, this link is mirrored.

For now we are not concerned with whether the activities should be
allowed to be linked, or about warning the user about implicitly
unlinking any previously associated activities when creating a new link
association.

Co-authored-by: Ynda <yndajas@gmail.com>