Allow linking ISPF ODA and non-ODA programmes
Only programmes of the opposite ODA type associated with the same
partner organisation can be linked.

Note: linking across different partner organisations would create too
much complexity at level C and D, where partner org users do not have
permissions to interact with other orgs’ activities. This is why we have
not included it for the MPV, since it is not required yet in practice
and refers to potential future use cases.

Co-authored-by: Ynda <yndajas@gmail.com>