Programmes can be linked through the UI
The first stage is enabling programmes to be linked via the interface.

The step is only enabled for ISPF programmes.

We’ve used the minimal implementation of radio buttons, which is enough
to allow us to develop the linking functionality and extend it to child
activities. Subsequent work will most likely change the radio boxes to
an accessible select with autocomplete functions.