Restrict linkable activities
In order to avoid accidentally breaking any existing links, do not allow
linking to activities that are already part of a pair of linked
activities.

In case any mistakes need to be fixed, users will have to manually
unlink the existing pair first.

We have discussed this in a design collaboration, and the participants
agreed that this situation occurs rarely enough that we can tolerate the
higher inconvenience.