Hide ISPF level B templates with feature flag
Hide the ISPF templates for level B bulk uploads when the ISPF feature
flag is enabled.

Level B activities are the only custom templates (level B budgets aren’t
fund specific). Level C uploads are only shown in the context of an
active report, and if the feature flag is enabled ISPF reports couldn’t
be created, or wouldn’t be shown.