Fix activity summary table layout
The add/edit link that was added to the activity summary was appearing to the
right of where the table was expected to end. This was not the case in the
activity details table. The difference between them was that in the details
table, every row had an actions cell regardless of whether it had content. This
brings the summary table in line, adding actions cells for all rows, even though
only one of them has content