Validate linked activities
This adds a custom validator to check if a given activity can be
linked to another activity

This will be used in particular to validate bulk uploaded activities

This commit also includes a bunch of test fixes that were needed after
adding the model validation

We also refactor the shared activity spec to no longer test all activity
levels here, as the test involved a fair amount of complex setup which
we don't think is necessary. We only need to test one of projects/third
party projects here.

Co-authored-by: Cristina <579522+CristinaRO@users.noreply.github.com>
Co-authored-by: George <george@dxw.com>