Reject ODA fields in non-ODA activity upload
If a user uploads a completed ODA bulk upload template using the non-ODA form, we
don't want to create a non-ODA activity with the additional fields, but instead
reject the upload since it doesn't match out expectations for ODA typing. This
safeguards against uploading activities as the wrong type by mistake. This only
seems to be possible in one direction, since the ODA templates have all the
non-ODA fields and more, while the non-ODA templates don't include various required
fields for ODA

This compares the ODA and non-ODA attributes and adds an error if any unexpected
attributes appear in the non-ODA upload