ISPF activities can have no ISPF partner country
Allow users to designate an ISPF activity as having no ISPF partner
countries. This option must be explicitly selected in the form, or
included in the uploaded data as `NONE`. This option is mutually
exclusive with selecting any country codes.

BEIS will write guidance to make it clear to partners what qualifies as
an exemption.