Move generic specs to the ODA context
The ODA context fulfills the “no linked activity” description too, but
the relevant setup is that it’s an ODA ISPF project, and is therefore
expected to have all the usual basic and project details, as well as
ISPF-specific details.

The non-ODA context needs to test the presence of ISPF-specific details
too, but not the “basic” and “project” details, which have ODA-specific
fields.

The “no linked activity” context is set up as a contrast to “already has
a linked activity” context because the specific thing we test is the
text of the action for the linked activity, adding or editing. (This is
true for both ODA and non-ODA projects, so we only tested one.)