Add organisation-less tests for `SpendingBreakdown`
Instances of `SpendingBreakdown` can be initialised without an organisation,
which changes the activities included in the export. This is used in the case of
downloading a spending breakdown from `/exports` rather than via an organisation's
exports page. This context was not previously tested anywhere in this spec, so
we've added a context with a couple of examples to cover this

Co-authored-by: George <george@dxw.com>