Extract `activity/import_spec.rb` class methods
This is a temporary change in the process of restructuring/rehousing the tests
in the `import_spec` and `import_level_b_spec`. For now, I've created a
`new_import_spec`, where I'll move all the tests that should stay in the
`import_spec`, and then tests that should be in new, different specs will be
moved where they should belong. The plan is to empty the `import_spec` and
`import_level_b_spec` files entirely, delete them and then rename
`new_import_spec` `import_spec`, so it becomes the new, sole, correctly-named
`import_spec`