Replace test for updating non-existent activity
The existing `Activity::Import` spec had a test for trying to update an
activity that doesn't exist. The handling of this specific error is the
responsibility of `ActivityUpdater` - `Import` shouldn't care about
specific errors, only whether or not there are errors. Therefore,
an `ActivityUpdater` test has been added to cover this specific error

Meanwhile, in `new_import_spec` we have a new `describe` block testing
the `import` method. There are two naive tests to confirm that it acts
per its definition: calling `import_row` once per row, and rolling
back any changes if there are errors. There's also a test that doesn't
strictly adhere to the single responsibility principle, checking that
the method successfully creates and updates activities with valid data.
Given that the `import_row` method is now private (not that it had any
specific tests before anyway), this feels like important coverage,
although we may still be doing/duplicating the work of feature tests.
Perhaps the test that it calls `import_row` once per row is enough?

This also makes minor changes to the actual `Import` and
`ActivityUpdater` classes to make a few methods private and use an
existing `attr_reader`