Fix PO identifier validation and add tests
Previously `Activity::Import` was meant to check that when updating,
the partner organisation identifier was not present - presumably it's
not meant to be updatable. However, it was checking the heading
"Partner Organisation Identifier", when the template's heading (and test
object's key) was "Partner organisation identifier". The case mismatch
meant that this validation wasn't working properly - it would always see
`nil` because it was looking for a heading that didn't exist. This fixes
the validation and updates the test specs not to try to update the
field

This was not covered in the test suite, so a test has been added to
check that the error shows up properly - this is in a new section of the
new import spec covering all the pre-action (pre-create/pre-update)
validation checks in `Activity::Import`. Errors are set within private
methods, but it feels important to check them nonetheless