Disable asset debugging mode in development
This was outputting an `application.debug-[fingerprint].js` file, which
was causing a lot of confusion. [This
issue](https://github.com/rails/jsbundling-rails/issues/73) had the
solution for why it was happening. I don't think there's much benefit to
outputting the `.debug` file if we're going to minify the JS (coming in
the next commit), so I've turned it off.