Minify JS with terser package
I've had to disable sourcemaps here for the time being, as I can't seem
to get them working and building with Yarn was throwing an error. As we
have fairly small amounts of JS in the code, I don't think we need to
worry too much about this now, but it would be good to revisit this in
future.