Exclude programmes from IATI exports when requested
This should probably have been included in 82476f3 (PR #1870), where the
option to flag programmes for exclusion from IATI was added. An
assumption was made there that that was all that was needed to exclude
them, however it appears that a scope was individually
applied to projects and third-party projects, and an equivalent was not
done for programmes. This applies the controller and spec pattern used
for projects and third-party projects to programmes