Reflect adjustments/refunds in level B IATI export
For programmes, BEIS would expect the transactions in the IATI export to
be the net spend, i.e. the total of actuals by financial quarter less
adjustments and refunds for that same quarter. This updates the
appropriate method to do this

For projects and third-party projects, the method just returns all of their
actuals. This is expected behaviour, but we will look into including adjustments
and probably refunds at those levels in a later commit (and PR)