Fix logout issues
37ce062 and fb834b8 from #1937 removed `@rails/ujs` in the process of moving to
js-bundling. The logout link is created using the `link_to` helper but with
`method :delete`. `a` tags with a `DELETE` method are not natively supported by
browsers - the way it worked before was that `@rails/ujs` was intercepting link
click events and making `DELETE` requests. Removing `@rails/ujs` has therefore
broken the logout link

This brings the package back and loads it in `application.js`. Previously, and
conventionally, when including `Rails` from `@rails/ujs`, somewhere in the same
file `Rails.start()` is called. However, this doesn't appear to be necessary to
restore the logout functionality

Alternatively, we could change the Devise initializer/config to allow logout
requests to be processed via `GET`, but this might be more unconventional

The `users_can_sign_out` spec didn't catch this, but it can be made to fail (when
`@rails/ujs` is not installed) by adding `js: true` to the relevant test. This
has now been done to guard against future regression