Fix inconsistent page/question titles
On the activity form, some page titles were not in sync with the
wording of the question. This brings them in line, while also
substantially DRYing up the activity translations file

In the objectives form step, the `fieldset` is removed since there's only one
input. Previously there was a `fieldset` legend, which was the full question,
then a label on the `textarea` input which didn't add anything useful
("Aims/Objectives of the activity"). The label has therefore been updated to the
full question, with all else removed