Refactor linked activity translations
This brings linked activity translations more in line with those of other fields

Instead of a separate condition in the `ActivityFormsController` `page_title`
method, the translations have been refactored to work in the same way as others,
with the `linked_activity_type` passed in as an argument to `I18n.t`

The hint translation has also had the `_guidance` suffix removed, since this is
implied by being a `hint` (other hints don't use it)

In the `ActivityFormsContoller`'s `page_title` method, various arguments are
passed into the `I18n.t` call, but these are simply ignored when the translation
doesn't need them. We send all three so that each is available when a given
translation needs it (e.g. `linked_activity_type` is needed in the
`linked_activity` translation, but including it in calls for other translations
has no side effect). This was already the appraoch before this commit - this
commit simply adds another argument and some line breaks for readability