Remove redundant activity translations
Following a refactor of the activity form translations, this removes
remaining translations that appear redundant. These are mostly from old
fields that are no longer in use, but there are also some translations
for remaining fields that aren't used in their respective areas (e.g.
where codelists are used in their place)