Upload approved report CSV to S3
Upload the report CSV to S3 when the report is approved. The purpose of this feature is to provide BEIS a snapshot of the activities included in the report as they were when the report was approved.

When the report is generated from live data, it can also contain later changes to activity attributes, or adjustments from later reports. Uploading the file will allow RODA to provide the users with the information frozen in time, by replacing the live-generated download with the stored one.

Co-authored-by: George <george@dxw.com>
Co-authored-by: Ynda <yndajas@gmail.com>