Generate download of stored report CSV
Unlike the spending breakdown, which we have been emulating so far,
reports contain information that shouldn’t be public by default.

Rather than exposing a public URL, and given the constraints around
generating signed URLs for buckets created via Terraform[1], we have
chosen the approach of generating the download on the spot and streaming
it to the user’s browser.

This way there is no public URL, and the download link is secured behind
RODA authentication and authorisation.

Generating the download from the RODA app also requires giving the app
read permissions for the s3 bucket (previously only the worker process
needed to be able to exchange data with the s3 storage).

[1] https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1678