Test: make s3 export bucket private
We'd like to verify if making this bucket private works properly on
staging with this terraform change. This will likely break the spending
breakdown functionality, as users will be unable to access the report
via its public url.

We'll be using this to securely store an organisation's report CSVs to
ensure the data stays the same once approved.