Prefix all non-ODA identifiers with “NODA”
BEIS users need a way to tell quickly, at a glance, whether an activity
is ODA or non-ODA. We discussed various options, including whether to
add something to ODA activities as well, and in the end we settled on
adding the prefix to non-ODA activities only.

There was also a discussion about where this identifier fragment should
go, and Josh expressed a preference for it as a prefix rather than
inserted after the fund identifier.