Register custom driver to run JavaScript tests
Following an update to the latest version of `selenium-webdriver`, the default
`selenium_headless` driver shows a deprecation warning when running the specs,
as it uses the `capabilities` key when registering the driver.

`WARN Selenium [DEPRECATION] [:capabilities] The :capabilities parameter for
Selenium::WebDriver::Firefox::Driver is deprecated. Use :options argument with
an instance of Selenium::WebDriver::Firefox::Driver instead.`

To remove the warning, register a custom driver so that we have greater control
over the options passed in. Otherwise, this is set up the same as the default
driver.