Fix edit render after failed comment update
Previously if you tried to update a comment and it was invalid (e.g. if
it had no body), the server would attempt to render the edit form but
the controller hadn't set the prerequisite `@activity` instance
variable. This sets that and adds a test to ensure the edit form is
rendered with an error message (and the comment is not updated)