Use RODA ID in absence of title on comment form
If trying to add comments to an activity without a title, the forms
were previously just headed "Add a comment about" or "Edit comment
about". The title is meant to follow these, but if it doesn't exist,
they read a little confusingly

This makes it so that the RODA identifier is used in place of the title
in such cases

This is likely an extreme edge case. The likelihood of needing to add
a comment to an incomplete activity (especially so incomplete as not to
have a title) is surely low, so perhaps the tests are unnecessary, but
I thought it was worth cleaning up alongside other comment-related fixes