Use the private bucket for spending breakdown
It was previously agreed by BEIS that spending breakdown data is data that would be public anyway, as part of the duty to be transparent, which meant that we were comfortable using a public S3 bucket. (The file URLs were not guessable, nor publicly listed.)

In recent discussions, occasioned by report CSVs being uploaded to S3, the BEIS team have pointed out that the data is not necessarily public immediately, but after a process of review and approval. BEIS decided that as a matter of principle this data shouldn’t be in the public space, so we are switching to using the same private S3 bucket that we’re using for report CSVs.