Show download for existing spending breakdown files
This makes the latest generated spending breakdown export available through RODA. Users can always request a new export to be generated.

In a future commit we’ll be removing the sharing of a public URL for spending breakdown exports, so the exports page will become the only place to obtain the downloads.