Update ruby to 2.7.7
Support for ruby 2.x ends on 31 March 2023.

Brakeman warnings are making the builds fail. We want to unblock our
builds while we update to ruby 3 and iron out any snags.

We’re also updating to ruby 2.7.7 (same EOL) to benefit from the
security update in this last supported 2.x version.