Fix `rows_for_last_financial_quarter`
A while back, we added protection to `rows_for_first_financial_quarter`
to prevent it trying to convert `nil` values to BigDecimals when there
weren't any found.

Although [the
PR](https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/1482)
in which that was added suggests that there should always be a value for
the last row of actuals, on Staging, at least, this doesn't appear to be
the case.

Although the above may be due to slightly wonky, possibly invalid data
on Staging, this error caused a report (with ID
b45c5b36-e502-4406-ac66-8a51321a177c) to fail being uploaded to S3. We'd
like to make sure this doesn't happen on Production.