Add migration to add Chile as a benefitting country to old activities
We have received a support query regarding two old activities that should have
Chile as a benefitting country, but don't.

Chile is no longer available in the UI as it is no longer an eligible
benefitting country, however, it was at the time the activities were active.

Add a data migration that adds Chile as a benefitting country to the activities
in question.