Optimise Codelist.new in development
In 50088fa, Stu suggested it's potentially useful not to memoise this
method in development, but this led us down a garden path when debugging
an IATI export timeout issue. What we thought was a big optimisation
ended up being specific to the development environment due to this
conditional

It's easy enough to reload your local server in the very occasional
times that we make changes to codelists. It's preferable to keep
the development environment as production-like as possible in this case,
in order to speed things up and aid any future debugging work

Co-authored-by: Cristina <579522+CristinaRO@users.noreply.github.com>
Co-authored-by: George <george@dxw.com>
Co-authored-by: Joe <jkempster34@gmail.com>
Co-authored-by: Suze <suze@dxw.com>