Clear remember token during training DB sync
After running the training database sync script, some users have
reported being unable to log in, encountering an `Internal Server
Error` screen and/or being unable to reset their password
successfully. The latter will be fixed in the next commit

When users encounter the internal server error message, the
corresponding server error is as follows:

```
RuntimeError: authenticatable_salt returned nil for the User model.
In order to use rememberable, you must ensure a password is always
set or have a remember_token column in your model or implement your
own rememberable_value in the model with custom logic.
```

Using a different browser or an incognito window, or deleting the
`_roda_session` cookie all seemed to get around the error, which
led us to think it might be to do with sessions. However, clearing
sessions on the server did not fix this error

Given the error message mentions a `remember_token`, we next
tried using Devise's `forget_me!` method on users during the
`force_password_reset_for_users` method/step, which removes the
token if it exists

This successfully fixed the internal server error for @CristinaRO,
who was experiencing it while work on this solution was underway