Fix deprecated keyword argument behaviour
Ruby 2.7 deprecated certain behaviours of keyword arguments; these will be a
breaking change in Ruby 3.0. 

Before updating to Ruby 3.0, fix the deprecation warnings.

Some forms were using `include_blank` incorrectly, which meant that the select
component of the formbuilder wasn't generating the desired blank element. Now
that this is fixed, we can remove the structs that were replicating this
behaviour in the FormHelper.