Fix `CSV#open` args in ReportExportUploaderJob
and test outcome of "real" ReportExportUploderJob.

When we updated to Ruby 3, our tests didn't catch an error
that came from the way Ruby now handles keyword args passed
to its method as a hash, which caused an error to be thrown
when calling `CSV#open` in the `ReportExportUploader` class.

We're using a regex here to check the uploaded filename, rather than
stubbing the date, as that would require mocking various `Time`s that
pass validations e.g. not being in the future, which would be fairly
complex to set up and likely brittle.