Make RODA ID spec more robust
This was using built rather than created activities to get a parent RODA
identifier, but build activities don't have them. It was then checking
that the RODA identifier of a child activity included the parent's RODA
identifier - "" - which technically it did.

This refactor mocks the `roda_identifier` method of the parent
activities and then explicitly checks that they appear in the child RODA
identifiers and not as an empty string.

This was surfaced by adding `.compact` to implement the conditional
`NODA-` prefix (coming in a future commit `ee6dfa9`) which broke a test
because without `.compact` the RODA identifiers did indeed start "-",
whereas with `.compact` that first (`.join`-originating) dash was gone.

Co-authored-by: Ynda Jas <ynda@dxw.com>