Imported forecasts
If the values are included in the CSV being uploaded, list them as
imported even if they are the same as the existing values (which means
nothing was imported per se).

Do not list the deleted values when they have been replaced by zero
forecasts. This is an improvement over listing the now-deleted values.
It doesn’t address the users’ need to see explicitly that the forecast
value has been replaced with zero, because in the case of original
forecasts (which didn’t have any other values reported in previous
reports), we do not replace with zero, but delete the original as to not
create a zero-value original “revision”.