Clarify behaviour
The behaviour might be unusual, so it needs documenting by writing a
spec for it.

Returning the value of a deleted entity is not in itself an unusual
pattern, when the method is an explicit “delete”/“destroy” method.

However, `set_value` is not a delete method, which only incidentally has
the side effect of deleting an entry sometimes.

We could change the return value of the method in future, and
documenting it via a spec will remind us to mind any side effects of
changing the method’s behaviour.