Skip Pundit’s policy scope verification on index
In our base controller, we use Pundit’s mechanism for alerting us if we
forget to authorize any controller actions. Collective controller
actions such as index are checked for the use of `policy_scope`.

The reports index doesn’t call the policy scope, so we see this error in
our server logs:
`Pundit::PolicyScopingNotPerformedError (ReportsController)`

We can skip the policy scope check, because permissions are checked
against the user’s organisation.